chore: move function to its own file
